+++ src/main/java/net/geforcemods/securitycraft/blockentities/IMSBlockEntity.java
@@ -102,7 +102,7 @@ public class IMSBlockEntity extends CustomizableBlockEntity implements ITickingB
 	}
 
 	public boolean canAttackEntity(LivingEntity entity) {
-		return entity != null && (!(entity instanceof Player player) || !(isOwnedBy(player) && ignoresOwner()) && !player.isCreative() && !player.isSpectator()) //Player checks
+		return entity != null && (!(entity instanceof Player player) || !(isOwnedBy(player) && ignoresOwner()) && player.canBeSeenAsEnemy()) //Player checks
 				&& !(isAllowed(entity)); //checks for all entities
 	}