+++ src/main/java/net/geforcemods/securitycraft/blocks/mines/BouncingBettyBlock.java
@@ -10,6 +10,7 @@ import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.entity.Entity;
+import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.BlockGetter;
@@ -58,7 +59,7 @@ public class BouncingBettyBlock extends ExplosiveBlock {
 	public void entityInside(BlockState state, Level level, BlockPos pos, Entity entity) {
 		if (!getShape(state, level, pos, CollisionContext.of(entity)).bounds().move(pos).inflate(0.01D).intersects(entity.getBoundingBox()))
 			return;
-		else if (!EntityUtils.doesEntityOwn(entity, level, pos))
+		else if (!EntityUtils.doesEntityOwn(entity, level, pos) && (!(entity instanceof LivingEntity livingEntity) || livingEntity.canBeSeenAsEnemy()))
 			explode(level, pos);
 	}