+++ src/main/java/net/geforcemods/securitycraft/blocks/mines/MineBlock.java
@@ -8,6 +8,7 @@ import net.geforcemods.securitycraft.util.EntityUtils;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
 import net.minecraft.world.entity.Entity;
+import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.item.ItemEntity;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
@@ -74,7 +75,7 @@ public class MineBlock extends ExplosiveBlock {
 			return;
 		else if (!getShape(state, level, pos, CollisionContext.of(entity)).bounds().move(pos).inflate(0.01D).intersects(entity.getBoundingBox()))
 			return;
-		else if (!EntityUtils.doesEntityOwn(entity, level, pos))
+		else if (!EntityUtils.doesEntityOwn(entity, level, pos) && (!(entity instanceof LivingEntity livingEntity) || livingEntity.canBeSeenAsEnemy()))
 			explode(level, pos);
 	}