+++ Changelog.md
@@ -39,6 +39,7 @@
 - Fix: Issues when unloading and saving chunks containing linkable block entities
 - Fix: Some scroll lists can be scrolled a tiny bit even though there is no scrollbar
 - Fix: Some scroll lists cannot be scrolled far enough to show all entries
+- Fix: Some mines explode when detecting a player in creative mode
 - Misc.: The minimum required Forge version is now 44.0.11
 
 --------------------------Changelog for v1.9.4-beta1 of SecurityCraft--------------------------